Refactoring the md code
We now the ability to pause and un-paused the sim. Through boundary collision detecting has been tuned off due to performance issues. Code is still running much slower than the original collision detection algorithm so there is something going on in terms of object creating that it surprisingly expensive. Added system state and time panel to bottom corner. It's not pretty but will aid with debugging and can be turned into something more aesthetically pleasing at a later date.