Major re-structuring: initial stage
As a test, the structure of the simple reactor application has been
augmented to make it more self-contained. Here 'self-contained' is
used in the sense that each app/* directory contains all the js, html,
css and image files required to run that app, above those present in
the top-level lib. This will hopefully make the application easier to
replicated. 'modules' has been renamed to 'apps' to reflect that these
are now intended to be independent applications. Some existing app
tests are likely to have been broken. Fixes to come.